Fixes Latest Version Downloading For PaperSpigot (#29)
* Fixes Latest Version Downloading

The original download link downloaded the 1.12.2 version instead of the latest. I fixed the link, however it will likely need to be watched for future changes.

* Updated to support new versions.

Removed the uncessesary version field, as it did nothing. Fixed the 1.13 link to reflect the current one. Added a 1.14 version file. Renamed the files to reflect the current stability of each.

* Updated to clarify naming.